+++ xdev/bin/freshpyenv.sh
@@ -31,9 +31,9 @@ OPTIONS
             view this help message
 
 EXAMPLES
-    ~/code/xdev/xdev/bin/freshpyenv.sh images
-    ~/code/xdev/xdev/bin/freshpyenv.sh --help
-    ~/code/xdev/xdev/bin/freshpyenv.sh --image python:3.10
+    freshpyenv.sh images
+    freshpyenv.sh --help
+    freshpyenv.sh --image python:3.10
 
 AUTHOR
      Jon Crall <erotemic@gmail.com>