+++ CHANGELOG.md
@@ -4,7 +4,21 @@ We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 1.3.4 - Released 2023-06-14 
+## Version 1.5.0 - Unreleased
+
+
+### Added
+
+* Add `xdev.snapshot`
+* Add attributes `file_paths` and `dir_paths` to `DirectoryWalker`.
+* Add `ignore_empty_dirs` argument to `DirectoryWalker`.
+
+### Fixed:
+
+* Fixed issue in `DirectoryWalker` with rich links.
+
+
+## Version 1.4.0 - Released 2023-08-18 
 
 ### Added
 * Add dirstats CLI